+++ tests/regression/transaction/tests.py
@@ -1,7 +1,7 @@
 import random
 
 import stdnet
-from stdnet import test, orm
+from stdnet import test, orm, InvalidTransaction
 from examples.models import SimpleModel, Dictionary
 from stdnet.utils import populate
 
@@ -9,13 +9,17 @@ LEN = 100
 names = populate('string',LEN, min_len = 5, max_len = 20)
 
 
-class TestTransactions(test.TestCase):
-    
-    def setUp(self):
-        self.orm.register(SimpleModel)
-    
-    def unregister(self):
-        self.orm.unregister(SimpleModel)
+class TransactionReceiver(object):
+    requests = {}
+    def __call__(self, sender, response = None, raw_command = None, **kwargs):
+        if raw_command in self.requests:
+            self.requests[raw_command]['response'] = response
+        else:
+            self.requests[raw_command] = kwargs
+        
+
+class TestTransactions(test.TestModelBase):
+    model = SimpleModel
         
     def testSave(self):
         with SimpleModel.transaction() as t:
@@ -38,11 +42,44 @@ class TestTransactions(test.TestCase):
         self.assertRaises(SimpleModel.DoesNotExist,
                           SimpleModel.objects.get,id=s.id)
         
-    
-class TestTransactionMultiFields(test.TestCase):
-    pass
-    
+    def testIncompatibleModels(self):
+        l = stdnet.struct.list('redis://localhost:6379/?db=9')
+        s = stdnet.struct.set('redis://localhost:6379/?db=8')
+        self.assertRaises(InvalidTransaction,orm.transaction,l,s)
+        l = stdnet.struct.list('redis://localhost:6379/?db=9')
+        s = stdnet.struct.set('redis://localhost:6378/?db=9')
+        self.assertRaises(InvalidTransaction,orm.transaction,l,s)
+        self.assertRaises(InvalidTransaction,orm.transaction,SimpleModel,s)
         
+    def testCompatibleModels(self):
+        s = stdnet.struct.set(SimpleModel._meta.cursor)
+        t = orm.transaction(SimpleModel,s, name = 'test-transaction')
+        self.assertEqual(t.server,SimpleModel._meta.cursor)
+        self.assertEqual(t.name,'test-transaction')
+        
+    def testSingleTransaction(self):
+        db = SimpleModel._meta.cursor
+        s = stdnet.struct.set(db)
+        l = stdnet.struct.list(db)
+        h = stdnet.struct.hash(db)
+        r = TransactionReceiver()
+        db.redispy.signal_on_send.connect(r)
+        db.redispy.signal_on_received.connect(r)
+        with orm.transaction(s,l,h,SimpleModel) as t:
+            m = SimpleModel(code = 'test', description = 'just a test')
+            h.add('test','bla',t)
+            l.push_back(5,t)
+            l.push_back(8,t)
+            s.update((2,3,4,5,6,7),t)
+            s.save(t)
+            l.save(t)
+            h.save(t)
+            m.save(t)
+        self.assertEqual(s.size(),6)
+        self.assertEqual(l.size(),2)
+        self.assertEqual(h.size(),1)
+        self.assertEqual(SimpleModel.objects.all().count(),1)
+        self.assertEqual(len(r.requests),1)
         
 class TestMultiFieldTransaction(test.TestModelBase):
     model = Dictionary
@@ -71,6 +108,8 @@ class TestMultiFieldTransaction(test.TestModelBase):
             self.assertTrue(len(t._cachepipes),2)
             for c in t._cachepipes:
                 self.assertEqual(len(t._cachepipes[c].pipe),2)
+                
+                
             
         self.assertTrue(len(t._cachepipes),2)
         for c in t._cachepipes: